backend/requirements.txt
huggingface-hub==0.19.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
fastapi==0.128.0
filelock==3.20.3
fsspec==2026.1.0

h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1

idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0

langsmith==0.6.6
numpy<2
orjson==3.11.5
packaging==24.2
pinecone==8.0.0
pinecone-plugin-assistant==3.0.1
pinecone-plugin-interface==0.0.7
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-settings>=2.10.1,<3.0.0
pypdf==6.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
rank-bm25==0.2.2
requests==2.32.5
requests-toolbelt==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
sentence-transformers==2.2.2
transformers==4.30.2
tenacity==9.1.2
tqdm==4.67.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.40.0
watchfiles==1.1.1
websockets==16.0
zstandard==0.25.0
